axio config: support non-HTTPS cookies
This change requires all axio-alpha clients to update their
configurations for axio-config (pass is-https?).

Presumably this configuration bit could be used by other components
where necessary, but the central use of axio-config-is-https? right now
is in create-session-cookie, where it is passed to make-id-cookie: the
previous defaults turn out to work in Chrome over HTTP but not Safari
(versions may matter).

This configuration provides a hook for developers to request HTTP during
local development of an axio app but for the deployed application to use
HTTPS (the default).

Also note a TODO: log-level is not currently exposed by axio-app-init,
which may be an oversight. This means log-level is not currently
configurable in the normal way (though you could probably still glue the
pieces together yourself).