url-for: support #:query params
The input values are the same as for the query part of the url struct
(an association list of symbols to strings or #f); fittingly, the
alist->form-urlencoded procedure does all the hard work for us. Since
this input contract is non-obvious, I've attached a contract to the
generated url-for function documenting what query-params should be; this
also improves error messages for wrong uses.

While we're at it clean up the require clauses.