syntax: make generate-url-for route-name optional
This means we no longer have to pass a dummy #f value in routes and we
avoid generating code like

    (define hsh (make-immutable-hash '((#f . "path1") (#f . "path2"))))

which is ambiguous (except that Racket documents that later keys
override earlier ones) and not useful (you're not expected to use the
url-for-fun with #f as the route-name).

A side benefit is we can ascribe the :id syntax-class to route-name.

The ~? form is a kind of conditional templating: in (~? t1 t2), if any
pattern variables in t1 are unbound, it expands to t2 instead. The
degenerate case of expanding t2 to _nothing_ is usually written (~? t1).
So you can read

    #'(f (path (~? route-name)) ...)

as a call to `f` with `[path]` when it has no route name and `[path
name]` when it does. Similarly for

    #'(define hsh (make-immutable-hash '((~? (route-name . path)) ...)))

For any instance of the ellipsis-marked pattern `(path:string (~optional
route-name:id))` where `route-name` is not present, _omit generating the
key-value pair for `make-immutable-hash`_.